requirements.txt
paddlepaddle>=2.5.0
paddleocr>=2.7.0
Pillow>=9.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
flask>=2.3.0
huggingface_hub>=0.19.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0